pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "vagent"
version = "0.1.0"
description = "Vertex AI agentic REPL with local tool execution"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "google-auth>=2.0.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.scripts]
vagent = "vagent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
